fix unsigned long on unsigned long long

+++ include/task3.h
@@ -1,2 +1,2 @@
 #pragma once
-unsigned long sumPrime(unsigned int hbound);
+unsigned long long sumPrime(unsigned int hbound);
