+++ src/task3.cpp
@@ -2,9 +2,9 @@
 #include "task2.h"
 #include <cmath>
 
-unsigned long sumPrime(unsigned int hbound)
+unsigned long long sumPrime(unsigned int hbound)
 {
-	unsigned int result{ 0 };
+	unsigned long long result{ 0 };
 	for (int i = 2; i < hbound; i++)
 		if (checkPrime(i))
 			result += i;
